backend/requirements.txt
# API Framework
fastapi==0.100.0
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
motor==3.2.0
pydantic==2.0.3
pymongo==4.4.1
beanie==1.21.0

# Resume Parsing
python-docx==0.8.11
PyPDF2==3.0.1
pdfminer.six==20221105

# NLP and ML
spacy==3.6.1
transformers==4.30.2
sentence-transformers==2.2.2
nltk==3.8.1
scikit-learn==1.3.0

# Utilities
python-jose==3.3.0  # For JWT token handling
passlib==1.7.4      # For password hashing
bcrypt==4.0.1       # For password encryption 